Do internal file structure refactor
This is an internal refactor to split the large `LgTvManager`
implementation across multiple files. Some other files are also renamed
for clarity.

The external API is unchanged.
